+++ Source/tinyxml/byte_stream.h
@@ -37,7 +37,7 @@ namespace TinyXPath
 class byte_stream
 {
    /// Length of the total std::string, + 1
-   unsigned u_length;
+   size_t u_length;
    /// Total std::string
    _byte_ * bp_in;
    /// Current read position
